Added Login Services

+++ src/main/java/com/jeido/exercisespring/controllers/StudentController.java
@@ -1,6 +1,7 @@
 package com.jeido.exercisespring.controllers;
 
 import com.jeido.exercisespring.entities.Student;
+import com.jeido.exercisespring.services.LoginService;
 import com.jeido.exercisespring.services.StudentService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,13 +18,20 @@ public class StudentController {
 
     private final StudentService studentService;
 
+    private final LoginService loginService;
+
+
     @Autowired
-    public StudentController(StudentService studentService) {
+    public StudentController(StudentService studentService, LoginService loginService) {
         this.studentService = studentService;
+        this.loginService = loginService;
     }
 
     @RequestMapping("/student/register")
     public String register(Model model) {
+        if (!loginService.isLoggedIn()) {
+            return "redirect:/login";
+        }
         Student student = Student.builder().name("").surname("").age(0).email("").build();
         model.addAttribute("student", student);
         model.addAttribute("action", "/student/register");
@@ -33,11 +41,15 @@ public class StudentController {
 
     @PostMapping("/student/register")
     public String addStudent(@Valid @ModelAttribute("student") Student student, BindingResult bindingResult, Model model) {
+        if (!loginService.isLoggedIn()) {
+            return "redirect:/login";
+        }
         if (bindingResult.hasErrors()) {
             model.addAttribute("action", "/student/register");
             model.addAttribute("mode", "add");
             return "student/details";
         }
+
         Student savedStudent = studentService.save(student.getSurname(), student.getName(), student.getAge(), student.getEmail());
         model.addAttribute("student", savedStudent);
         model.addAttribute("action", "");
@@ -47,6 +59,9 @@ public class StudentController {
 
     @RequestMapping("/student/{id}")
     public String student(@PathVariable("id")UUID id, Model model) {
+        if (!loginService.isLoggedIn()) {
+            return "redirect:/login";
+        }
         Student student = studentService.findById(id);
         model.addAttribute("student", student);
         model.addAttribute("mode", "info");
@@ -77,12 +92,18 @@ public class StudentController {
 
     @GetMapping("/student/delete/{id}")
     public String deleteStudent(@PathVariable("id") UUID id) {
+        if (!loginService.isLoggedIn()) {
+            return "redirect:/login";
+        }
         studentService.delete(id);
         return "redirect:/student";
     }
 
     @GetMapping("/student/edit/{id}")
     public String editStudent(@PathVariable("id") UUID id, Model model) {
+        if (!loginService.isLoggedIn()) {
+            return "redirect:/login";
+        }
         Student student = studentService.findById(id);
         model.addAttribute("student", student);
         model.addAttribute("mode", "edit");
@@ -92,6 +113,9 @@ public class StudentController {
 
     @PostMapping("/student/edit/{id}")
     public String studentEdit(@PathVariable("id") UUID id,@Valid @ModelAttribute("student") Student student, BindingResult bindingResult, Model model) {
+        if (!loginService.isLoggedIn()) {
+            return "redirect:/login";
+        }
         if (bindingResult.hasErrors()) {
             model.addAttribute("mode", "edit");
             model.addAttribute("student", student);
@@ -103,6 +127,25 @@ public class StudentController {
         model.addAttribute("mode", "info");
         model.addAttribute("action", "");
         return "redirect:/student/" + id;
+    }
+
+    @RequestMapping("/login")
+    public String login() {
+        return "login/form";
+    }
+
+    @PostMapping("/login")
+    public String login(@RequestParam("username") String username, @RequestParam("password") String password) {
+        if (loginService.login(username, password)) {
+            return "redirect:/student";
+        }
+        return "login/form";
+    }
+
+    @RequestMapping("/logout")
+    public String logout() {
+        loginService.logout();
+        return "redirect:/student";
     }
